requirements.txt
tldextract
tiktoken
google-genai